refactor: update delete cmd to follow the common cmd pattern
