+++ pkg/cmd/generate/builders/appconfig_builder_test.go
@@ -1,19 +1,34 @@
+// Copyright 2024 KusionStack Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package builders
 
 import (
+	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/bytedance/mockey"
 	"github.com/stretchr/testify/assert"
 	"kcl-lang.io/kpm/pkg/api"
-	pkg "kcl-lang.io/kpm/pkg/package"
 
 	v1 "kusionstack.io/kusion/pkg/apis/api.kusion.io/v1"
 	internalv1 "kusionstack.io/kusion/pkg/apis/internal.kusion.io/v1"
 	"kusionstack.io/kusion/pkg/modules"
 )
 
-func TestAppsConfigBuilder_Build(t *testing.T) {
+func TestBuild(t *testing.T) {
 	p, s := buildMockProjectAndStack()
 	appName, app := buildMockApp()
 	acg := &AppsConfigBuilder{
@@ -23,15 +38,16 @@ func TestAppsConfigBuilder_Build(t *testing.T) {
 		Workspace: buildMockWorkspace(),
 	}
 
-	kpmMock := mockey.Mock((*api.KclPackage).GetDependenciesInModFile).Return(&pkg.Dependencies{Deps: make(map[string]pkg.Dependency)}).
-		Build()
 	callMock := mockey.Mock(modules.CallGenerators).Return(nil).Build()
 	defer func() {
-		kpmMock.UnPatch()
 		callMock.UnPatch()
 	}()
 
-	kclPkg := &api.KclPackage{}
+	cwd, _ := os.Getwd()
+	pkgPath := filepath.Join(cwd, "testdata")
+	kclPkg, err := api.GetKclPackage(pkgPath)
+	assert.NoError(t, err)
+
 	intent, err := acg.Build(kclPkg, p, s)
 	assert.NoError(t, err)
 	assert.NotNil(t, intent)
